feat(map): 图谱可读性优化——缩小标签、撑开布局、聚合关系边
- 节点标签字号 10→7，加白色描边与 min-zoomed-font-size，缩放时自动收字
- fcose 斥力 4500→24000、弹簧 40–280→80–420、加 gravity/nodeSeparation，节点分得更开
- 孤立论文从叠字竖列改为右侧 4 列网格
- 断言边默认不画文字标签，仅悬停/选中时显示关系名；悬停节点高亮其关系、淡化其余
- 同一对论文 × 同语义组的多条关系用 bundleEdges 合并成带 ×N 的边（每对≤3）
- cites 边改在 visibleMapEdges 过滤，合并后仍保持默认隐藏
- 修 lineage edge_len NUMERIC→float8 cast

+++ server/crates/worker/src/neighbors.rs
@@ -156,7 +156,7 @@ pub async fn update_lineage(ctx: &JobContext, job: &Job) -> anyhow::Result<()> {
         SELECT
             src.anchor_work_id AS a,
             tgt.anchor_work_id AS b,
-            CASE WHEN r.review_status = 'confirmed' THEN 1.0 ELSE 2.0 END AS edge_len
+            (CASE WHEN r.review_status = 'confirmed' THEN 1.0 ELSE 2.0 END)::float8 AS edge_len
         FROM relations r
         JOIN relation_members src ON src.relation_id = r.id AND src.role = 'source'
         JOIN relation_members tgt ON tgt.relation_id = r.id AND tgt.role = 'target'
